gemini-rsvp/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
pydantic==2.11.9
python-dotenv==1.1.1
google-generativeai==0.7.2